requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
APScheduler==3.10.4
PyJWT==2.8.0
cryptography==41.0.7
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
